Changes names of view models
