+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Options/ConfigPathViewModel.cs
@@ -1,0 +1,21 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace NuGet.PackageManagement.UI.Options
+{
+    public class ConfigPathViewModel
+    {
+        public string ConfigPath { get; set; }
+        public ConfigPathViewModel(string path)
+        {
+            ConfigPath = path ?? throw new ArgumentNullException(nameof(path));
+        }
+
+        public override string ToString()
+        {
+            return ConfigPath;
+        }
+    }
+}
